Beginning the merge
I've merged the stock and order apps and got the app to run again.
There's probably more problems, but I need to commit for dev night.
