+++ EVERoachCoach/settings.py.example
@@ -38,8 +38,7 @@ INSTALLED_APPS = [
     'django.contrib.sessions',
     'django.contrib.messages',
     'django.contrib.staticfiles',
-    'stock',
-    'orders',
+    'rapid_sales_tool',
     'portal',
 ]
