--- orders/apps.py
@@ -1,7 +1,0 @@
-from __future__ import unicode_literals
-
-from django.apps import AppConfig
-
-
-class OrdersConfig(AppConfig):
-    name = 'orders'
