+++ rapid_sales_tool/admin.py
