+++ rapid_sales_tool/apps.py
@@ -3,5 +3,5 @@ from __future__ import unicode_literals
 from django.apps import AppConfig
 
 
-class StockConfig(AppConfig):
-    name = 'stock'
+class OrdersConfig(AppConfig):
+    name = 'rapid_sales_tool'
